[refactor] SendQueue
