+++ src/adapters/custom.ts
@@ -36,7 +36,7 @@ export class CustomAdapter extends BaseAdapter {
     try {
       return {
         model: response.model,
-        created_at: response.created,
+        created: response.created,
         message: {
           role: response.choices[0].message.role,
           content: response.choices[0].message.content,
