+++ src/utils/content.ts
@@ -1,7 +1,7 @@
 import { Session } from 'koishi';
 
 import { getMemberName } from './prompt';
-import { replaceImageWith } from './imageViewer';
+import { replaceImageWith } from '../services/imageViewer';
 import { Config } from '../config';
 import { emojiManager } from '../managers/emojiManager';
 import { convertUrltoBase64 } from './imageUtils';
